+++ asusctl/src/aura_cli.rs
@@ -1,8 +1,8 @@
+use gumdrop::Options;
 use rog_types::{
-    aura_modes::{Colour, Direction, Speed, AuraEffect, AuraModeNum, AuraZone},
+    aura_modes::{AuraEffect, AuraModeNum, AuraZone, Colour, Direction, Speed},
     error::AuraError,
 };
-use gumdrop::Options;
 use std::str::FromStr;
 
 #[derive(Options)]
@@ -165,7 +165,6 @@ impl Default for SetAuraBuiltin {
     }
 }
 
-
 impl From<&SingleColour> for AuraEffect {
     fn from(aura: &SingleColour) -> Self {
         Self {
@@ -294,15 +293,15 @@ impl From<&SetAuraBuiltin> for Vec<AuraEffect> {
                 zones[1].mode = AuraModeNum::Static;
                 zones[1].zone = AuraZone::Two;
                 zones[1].colour1 = data.colour2;
-                
+
                 zones[2].mode = AuraModeNum::Static;
                 zones[2].zone = AuraZone::Three;
                 zones[2].colour1 = data.colour3;
-                
+
                 zones[3].mode = AuraModeNum::Static;
                 zones[3].zone = AuraZone::Four;
                 zones[3].colour1 = data.colour4;
-            },
+            }
             SetAuraBuiltin::MultiBreathe(data) => {
                 zones[0].mode = AuraModeNum::Breathe;
                 zones[0].zone = AuraZone::One;
@@ -313,18 +312,18 @@ impl From<&SetAuraBuiltin> for Vec<AuraEffect> {
                 zones[1].zone = AuraZone::Two;
                 zones[1].colour1 = data.colour2;
                 zones[1].speed = data.speed;
-                
+
                 zones[2].mode = AuraModeNum::Breathe;
                 zones[2].zone = AuraZone::Three;
                 zones[2].colour1 = data.colour3;
                 zones[2].speed = data.speed;
-                
+
                 zones[3].mode = AuraModeNum::Breathe;
                 zones[3].zone = AuraZone::Four;
                 zones[3].colour1 = data.colour4;
                 zones[3].speed = data.speed;
             }
-            _ => {},
+            _ => {}
         }
         zones
     }